meta: retire MylesBorins from committee
When I originally joined the committee I had less responsibilities
elsewhere in the project. With my new position as TSC Director I
do not believe that I can commit the time necessary to properly
engage as a member of the CommComm

On the bright side, I am pretty sure I'll still be working closely
with all of you as the TSC Director, and look forward to collaborating!