Remove individual membership director role and move williamkapke to emeritus
Reflection of the state of the world with regard to Individual Membership Directors no longer existing as a role. Additionally, per the CommComm charter I as chairperson reached out to William Kapke after a period of inactivity in the work of the Community Committee, and recieved no indication that he is  interested in participating.  As such, I am moving him to emeritus.